+++ src/button.v
@@ -8,19 +8,18 @@ import gx
 pub struct Button {
 	Component_A
 pub mut:
-	app                &Window
-	icon               int
-	click_event_fn     fn (mut Window, Button)
-	new_click_event_fn fn (voidptr, voidptr, voidptr)
-	need_pack          bool
-	extra              string
-	user_data          voidptr
-	override_bg        bool
-	override_bg_color  gx.Color
-	icon_width         int
-	icon_height        int
-	border_radius      int
-	area_filled        bool = true
+	app               &Window
+	icon              int
+	old_click_fn      fn (voidptr, voidptr, voidptr)
+	need_pack         bool
+	extra             string
+	user_data         voidptr
+	override_bg       bool
+	override_bg_color gx.Color
+	icon_width        int
+	icon_height       int
+	border_radius     int
+	area_filled       bool = true
 }
 
 [params]
@@ -33,38 +32,27 @@ pub struct ButtonConfig {
 	icon        int  = -1
 }
 
-[deprecated: 'Use button(ButtonConfig)']
-pub fn button_with_icon(icon int, conf ButtonConfig) &Button {
-	cfg := ButtonConfig{
-		...conf
-		icon: icon
-	}
-	return button(cfg)
+pub fn button(cfg ButtonConfig) &Button {
+	return Button.new(cfg)
 }
 
-pub fn Button.new(cfg ButtonConfig) &Button {
-	return button(cfg)
-}
-
-pub fn button(conf ButtonConfig) &Button {
+pub fn Button.new(cf ButtonConfig) &Button {
 	return &Button{
 		app: unsafe { nil }
-		text: conf.text
-		icon: conf.icon
-		x: conf.bounds.x
-		y: conf.bounds.y
-		width: conf.bounds.width
-		height: conf.bounds.height
-		click_event_fn: fn (mut win Window, a Button) {}
-		new_click_event_fn: unsafe { nil }
-		user_data: conf.user_data
-		need_pack: conf.should_pack
-		area_filled: conf.area_filled
+		text: cf.text
+		icon: cf.icon
+		x: cf.bounds.x
+		y: cf.bounds.y
+		width: cf.bounds.width
+		height: cf.bounds.height
+		old_click_fn: unsafe { nil }
+		user_data: cf.user_data
+		need_pack: cf.should_pack
+		area_filled: cf.area_filled
 	}
 }
 
-// Sets the contentAreaFilled property, whether to paint
-// See https://docs.oracle.com/javase/7/docs/api/javax/swing/AbstractButton.html#setContentAreaFilled(boolean)
+// https://docs.oracle.com/javase/7/docs/api/javax/swing/AbstractButton.html#setContentAreaFilled(boolean)
 pub fn (mut this Button) set_area_filled(val bool) {
 	this.area_filled = val
 }
@@ -79,15 +67,56 @@ pub fn (mut btn Button) draw(ctx &GraphicsContext) {
 		btn.app = ctx.win
 	}
 
-	btn.app.draw_button(btn.x, btn.y, btn.width, btn.height, mut btn)
+	if btn.need_pack {
+		btn.pack_do(ctx)
+	}
+
+	text := btn.text
+	size := ctx.text_width(text) / 2
+	sizh := ctx.line_height / 2 // ctx.text_height(text) / 2
+
+	// Handle click
+	if btn.is_mouse_rele {
+		btn.handle_legacy_click()
+		btn.is_mouse_rele = false
+	}
+
+	// Draw Button Background & Border
+	btn.draw_background(ctx)
+
+	if btn.width == 0 && btn.height == 0 {
+		btn.pack_do(ctx)
+		btn.need_pack = true
+	}
+
+	if btn.icon != -1 {
+		wid := if btn.icon_width > 0 { btn.icon_width } else { btn.width }
+		hei := if btn.icon_height > 0 { btn.icon_height } else { btn.height }
+		ctx.gg.draw_image_with_config(gg.DrawImageConfig{
+			img_id: btn.icon
+			img_rect: gg.Rect{
+				x: btn.x + (btn.width / 2) - (wid / 2)
+				y: btn.y + (btn.height / 2) - (hei / 2)
+				width: wid
+				height: hei
+			}
+		})
+		return
+	}
+
+	ctx.draw_text((btn.x + (btn.width / 2)) - size, btn.y + (btn.height / 2) - sizh, text,
+		ctx.font, gx.TextCfg{
+		size: ctx.win.font_size
+		color: ctx.theme.text_color
+	})
 }
 
 pub fn (mut btn Button) pack() {
 	btn.need_pack = true
 }
 
-pub fn (mut btn Button) pack_do() {
-	width := text_width(btn.app, btn.text) + 6
+pub fn (mut btn Button) pack_do(ctx &GraphicsContext) {
+	width := ctx.text_width(btn.text) + 6
 	btn.width = width
 	btn.height = 30
 	btn.need_pack = false
@@ -148,65 +177,32 @@ fn (this &Button) get_bg(is_hover bool) gx.Color {
 	return this.app.theme.button_bg_normal
 }
 
-fn (mut app Window) draw_button(x int, y int, width int, height int, mut btn Button) {
-	if btn.need_pack {
-		btn.pack_do()
+// Deprecated functions:
+
+[deprecated: 'Use Button.new(ButtonConfig)']
+pub fn button_with_icon(icon int, conf ButtonConfig) &Button {
+	cfg := ButtonConfig{
+		...conf
+		icon: icon
 	}
-
-	text := btn.text
-	size := text_width(app, text) / 2
-	sizh := text_height(app, text) / 2
-
-	// Handle click
-	if btn.is_mouse_rele {
-		if app.bar == unsafe { nil } || app.bar.tik > 90 {
-			btn.click_event_fn(mut app, *btn)
-			if !isnil(btn.new_click_event_fn) {
-				btn.new_click_event_fn(app, btn, btn.user_data)
-			}
-		}
-		btn.is_mouse_rele = false
-	}
-
-	// Draw Button Text
-	ctx := app.graphics_context
-
-	// Draw Button Background & Border
-	btn.draw_background(ctx)
-
-	if btn.width == 0 && btn.height == 0 {
-		btn.pack_do()
-		btn.need_pack = true
-	}
-
-	if btn.icon != -1 {
-		wid := if btn.icon_width > 0 { btn.icon_width } else { btn.width }
-		hei := if btn.icon_height > 0 { btn.icon_height } else { btn.height }
-		ctx.gg.draw_image_with_config(gg.DrawImageConfig{
-			img_id: btn.icon
-			img_rect: gg.Rect{
-				x: btn.x + (btn.width / 2) - (wid / 2)
-				y: btn.y + (btn.height / 2) - (hei / 2)
-				width: wid
-				height: hei
-			}
-		})
-		return
-	}
-
-	ctx.draw_text((x + (width / 2)) - size, y + (height / 2) - sizh, text, ctx.font, gx.TextCfg{
-		size: app.font_size
-		color: ctx.theme.text_color
-	})
+	return button(cfg)
 }
 
-[deprecated: 'use subscribe_event']
-pub fn (mut com Button) set_click(b fn (mut Window, Button)) {
-	com.click_event_fn = b
+[deprecated]
+fn (mut app Window) draw_button(x int, y int, width int, height int, mut btn Button) {
 }
 
 [deprecated: 'use subscribe_event']
 pub fn (mut com Button) set_click_fn(b fn (voidptr, voidptr, voidptr), extra_data voidptr) {
-	com.new_click_event_fn = b
+	com.old_click_fn = b
 	com.user_data = extra_data
+}
+
+fn (mut btn Button) handle_legacy_click() {
+	mut win := btn.app
+	if win.bar == unsafe { nil } || win.bar.tik > 90 {
+		if !isnil(btn.old_click_fn) {
+			btn.old_click_fn(win, btn, btn.user_data)
+		}
+	}
 }